requirements.txt
langchain-core
langchain-openai
langgraph
pinecone
openai
pydantic
pymysql
python-dotenv
streamlit
